Xce dev tk 2019 05 21 (#233)
* XChemCOOT - test only: enable only refinement of bound-state once compchem stage is reached
* XChemDB: make more verbose due to unexpected behaviour when updating depositTable for specific project
* XChemDB: changed single to double quotation marks (and vice versa) in update depositTable function to allow for example 5' in title
* XChemDB: remove hidden characters like \r and \n when updating depositTable
* v1.3.8.3 (22/05/2019) - XChemCOOT: xce messages are written to xce.log file; XChemCOOT: warning message added in case cootOut/Refine_<n> folder exists

+++ lib/XChemCoot.py
@@ -644,30 +644,30 @@ class GUI(object):
         hbox.pack_start(frame)
 
         # label section --> start
-        frame = gtk.Frame(label='Label')
-        vbox = gtk.VBox()
-        labels = self.db.get_labels_from_db()
-        if len(labels) > 5:
-            print '==> sorry, too many labels; cannot display them in panel'
-        labels = labels[:5]
-        # with radiobuttons, one of them needs to be always on
-        # but there will be cases when the user has not assigned a label yet
-        # hence, the not_shown button is not shown but gets active
-        # if the label has not been set yet
-        labels.append('not_shown')
-        for n, l in enumerate(labels):
-            print n,l
-            if n == 0:
-                new_button = gtk.RadioButton(None, l)
-            else:
-                new_button = gtk.RadioButton(new_button, l)
-            new_button.connect("toggled", self.label_button_clicked, l)
-            if not l == 'not_shown':
-                vbox.add(new_button)
-            self.label_button_list.append(new_button)
-        frame.add(vbox)
-        hbox.pack_start(frame)
-        # label section <-- end
+#        frame = gtk.Frame(label='Label')
+#        vbox = gtk.VBox()
+#        labels = self.db.get_labels_from_db()
+#        if len(labels) > 5:
+#            print '==> sorry, too many labels; cannot display them in panel'
+#        labels = labels[:5]
+#        # with radiobuttons, one of them needs to be always on
+#        # but there will be cases when the user has not assigned a label yet
+#        # hence, the not_shown button is not shown but gets active
+#        # if the label has not been set yet
+#        labels.append('not_shown')
+#        for n, l in enumerate(labels):
+#            print n,l
+#            if n == 0:
+#                new_button = gtk.RadioButton(None, l)
+#            else:
+#                new_button = gtk.RadioButton(new_button, l)
+#            new_button.connect("toggled", self.label_button_clicked, l)
+#            if not l == 'not_shown':
+#                vbox.add(new_button)
+#            self.label_button_list.append(new_button)
+#        frame.add(vbox)
+#        hbox.pack_start(frame)
+#        # label section <-- end
 
         outer_frame.add(hbox)
         self.vbox.pack_start(outer_frame)
@@ -755,13 +755,13 @@ class GUI(object):
             self.refinement_folder = str(self.Todo[self.index][4])
             self.refinement_outcome = str(self.Todo[self.index][5])
             self.label = self.db.get_label_of_sample(self.xtalID)
-            self.update_label_radiobutton()
+#            self.update_label_radiobutton()
             self.update_RefinementOutcome_radiobutton()
         if self.xtalID not in self.siteDict:  # i.e. we are not working with a PanDDA model
             self.ligand_confidence = str(self.Todo[self.index][6])
             self.update_LigandConfidence_radiobutton()
             self.label = self.db.get_label_of_sample(self.xtalID)
-            self.update_label_radiobutton()
+#            self.update_label_radiobutton()
 
         self.RefreshData()
 
@@ -945,18 +945,25 @@ class GUI(object):
                     coot.set_last_map_colour(0.74, 0.44, 0.02)
 
     def experiment_stage_button_clicked(self, widget, data=None):
+        if data == '4 - CompChem ready' and self.refinementProtocol.startswith('pandda'):
+            print "==> XCE: removing refine.pdb and linking refine_<n>.split.bound-state.pdb to refine.pdb"
+            os.chdir(os.path.join(self.project_directory,self.xtalID))
+            if os.path.realpath('refine.pdb').replace('.pdb', '.split.bound-state.pdb'):
+                newPDB = os.path.realpath('refine.pdb').replace('.pdb', '.split.bound-state.pdb')
+                os.system('/bin/rm refine.pdb')
+                os.system('ln -s .%s refine.pdb' %newPDB.replace(os.getcwd(),''))
+                self.refinementProtocolcheckbox.set_active(False)
+                self.refinementProtocol = 'refmac'
+            else:
+                print '==> XCE ERROR: cannot find refine.output.bound-state.pdb'
+                pass
         self.db_dict_mainTable['RefinementOutcome'] = data
         print '==> XCE: setting Refinement Outcome for ' + self.xtalID + ' to ' + str(
-            data) + ' in mainTable of datasource'
-#        self.db.update_data_source(self.xtalID, self.db_dict_mainTable)
-        self.db.create_or_remove_missing_records_in_depositTable(self.xce_logfile,self.xtalID,'ligand_bound',self.db_dict_mainTable)
-        if data == '4 - CompChem ready' and self.refinementProtocol.startswith('pandda'):
-            print "==> XCE: removing refine.pdb and setting linking refine.split.bound-state.pdb to refine.pdb"
-            os.chdir(os.path.join(self.project_directory,self.xtalID))
-            os.system('/bin/rm refine.pdb')
-            os.system('ln -s refine.split.bound-state.pdb refine.pdb')
-            self.refinementProtocolcheckbox.set_active(False)
-            self.refinementProtocol = 'refmac'
+                data) + ' in mainTable of datasource'
+            #        self.db.update_data_source(self.xtalID, self.db_dict_mainTable)
+        self.db.create_or_remove_missing_records_in_depositTable(self.xce_logfile, self.xtalID, 'ligand_bound',
+                                                                     self.db_dict_mainTable)
+
 
     def ligand_confidence_button_clicked(self, widget, data=None):
         print 'PANDDA_index', self.pandda_index
@@ -988,14 +995,14 @@ class GUI(object):
 #                widget.get_label()) + ' in mainTable of datasource'
 #            self.db.update_data_source(self.xtalID, self.db_dict_mainTable)
 
-    def label_button_clicked(self, widget, data=None):
-        print '............',data
-        if data == 'not_shown':
-            self.db.execute_statement("update mainTable set label=Null where CrystalName = '%s'" %self.xtalID)
-        else:
-            self.db_dict_mainTable['label'] = data
-            print '==> XCE: setting label for ' + self.xtalID + ' to ' + str(data) + ' in mainTable of datasource'
-            self.db.update_data_source(self.xtalID, self.db_dict_mainTable)
+#    def label_button_clicked(self, widget, data=None):
+#        print '............',data
+#        if data == 'not_shown':
+#            self.db.execute_statement("update mainTable set label=Null where CrystalName = '%s'" %self.xtalID)
+#        else:
+#            self.db_dict_mainTable['label'] = data
+#            print '==> XCE: setting label for ' + self.xtalID + ' to ' + str(data) + ' in mainTable of datasource'
+#            self.db.update_data_source(self.xtalID, self.db_dict_mainTable)
 
     def RefreshData(self):
         # reset spider plot image
@@ -1321,6 +1328,12 @@ class GUI(object):
                     coot.write_pdb_file(item,
                                         os.path.join(self.project_directory, self.xtalID, 'Refine_' + str(self.Serial),
                                                      'in.pdb'))
+                    break
+                elif coot.molecule_name(item).endswith('refine.split.bound-state.pdb'):
+                    coot.write_pdb_file(item,
+                                        os.path.join(self.project_directory, self.xtalID,
+                                                        'Refine_' + str(self.Serial),
+                                                        'in.pdb'))
                     break
                 elif coot.molecule_name(item).endswith('init.pdb'):
                     coot.write_pdb_file(item,
